XML interpretation Ready
I’ve added a way to read xmlFiles from the amazonDataFolder.
Fixed the xml, to comply with the standard (there were some illegal
characters in the files).

Added a collection of classes to handle the Movies, Users, and ratings,
and their relationships.

Added also a small text processor to process titles, which come a bit
raw.

It’s needed to add a jar file from apache to get the average.
Download from:
http://commons.apache.org/proper/commons-math/download_math.cgi